Refactor user settings layout (#18)
