+++ resources/views/auth/settings/_settings-nav.blade.php
@@ -1,0 +1,14 @@
+<div class="col-12 pb-4">
+    <ul class="nav-settings nav-overflow nav">
+        <li class="nav-item">
+            <a class="nav-link active" href="{{ route('account.settings.information') }}">
+                <x-heroicon-o-information-circle class="icon mr-1"/> {{ __('Informatie') }}
+            </a>
+        </li>
+        <li class="nav-item">
+            <a class="nav-link" href="{{ route('account.settings.security') }}">
+                <x-heroicon-o-key class="icon mr-1"/> {{ __('Beveiliging') }}
+            </a>
+        </li>
+    </ul>
+</div>
