+++ routes/web.php
@@ -2,6 +2,9 @@
 
 use App\Http\Controllers\Front\DashboardController;
 use App\Http\Controllers\Front\AccountController;
+use App\Http\Controllers\Front\Settings\DeleteController;
+use App\Http\Controllers\Front\Settings\InformationController;
+use App\Http\Controllers\Front\Settings\SecurityController;
 use App\Http\Controllers\Front\WelcomeController;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
@@ -21,12 +24,13 @@ Auth::routes();
 Route::get('/', WelcomeController::class)->name('welcome');
 
 Route::group(['middleware' => ['auth', 'forbid-banned-user']], static function () {
-    Route::delete('/Account-verwijderen', [AccountController::class, 'destroy'])->name('account.delete');
+    Route::delete('/Account-verwijderen', DeleteController::class)->name('account.delete');
 
     Route::group(['prefix' => 'account-instellngen'], static function (): void {
-        Route::get('/', [AccountController::class, 'index'])->name('account.settings');
-        Route::patch('/informatie', [AccountController::class, 'updateInformation'])->name('account.settings.information');
-        Route::patch('/security', [AccountController::class, 'updateSecurity'])->name('account.settings.security');
+        Route::get('/informatie', [InformationController::class, 'index'])->name('account.settings.information');
+        Route::patch('/informatie', [InformationController::class, 'update'])->name('account.settings.information');
+        Route::get('/security', [SecurityController::class, 'index'])->name('account.settings.security');
+        Route::patch('/security', [SecurityController::class, 'update'])->name('account.settings.security');
     });
 
    Route::get('/home', DashboardController::class)->name('home');
